requirements.txt
paho-mqtt==1.4.0
pyserial==3.5

